backend/pyproject.toml
[project]
name = "Lukovka"
version = "0.1.0"
description = "Сервис покупки топлева"
authors = [
    { name = "Данил Погонцев", email = "data.name.id@yandex.ru" },
    { name = "Сергей Андреев", email = "gray-advantage@yandex.ru" },
    { name = "Алина Баландинская", email = "linkinunn@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "orjson (>=3.10.15,<4.0.0)",
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.7"
mypy = "^1.15.0"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
src = ["backend/app"]
target-version = "py312"

[tool.ruff.lint]
preview = true

select = ["ALL"]
ignore = ["D", "DOC", "CPY", "RUF001", "RUF002", "RUF003", "PLC0415"]

[tool.ruff.lint.pylint]
max-args = 10


[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
enable_incomplete_feature = ["NewGenericSyntax"]
